Move PurchasesByOtherAdminsNotice to own component
