+++ client/me/purchases/purchases-list/index.tsx
@@ -13,7 +13,6 @@ import NoSitesMessage from 'calypso/components/empty-content/no-sites-message';
 import InlineSupportLink from 'calypso/components/inline-support-link';
 import Main from 'calypso/components/main';
 import NavigationHeader from 'calypso/components/navigation-header';
-import Notice from 'calypso/components/notice';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import TrackComponentView from 'calypso/lib/analytics/track-component-view';
 import { getPurchasesBySite, getSubscriptionsBySite } from 'calypso/lib/purchases';
@@ -41,6 +40,7 @@ import { getSiteId } from 'calypso/state/sites/selectors';
 import { AppState } from 'calypso/types';
 import MembershipSite from '../membership-site';
 import PurchasesSite from '../purchases-site';
+import { PurchasesByOtherAdminsNotice } from './purchases-by-other-admins-notice';
 import PurchasesListHeader from './purchases-list-header';
 
 export interface PurchasesListProps {
@@ -68,60 +68,6 @@ class PurchasesList extends Component<
 		}
 
 		return ! this.props.sites.length && ! this.props.subscriptions.length;
-	}
-
-	renderPurchasesByOtherAdminsNotice() {
-		const { sites, translate } = this.props;
-
-		/*
-		 * Because this is only rendered when the user has no purchases,
-		 * we don't need to check each site to ensure the purchases aren't
-		 * linked with the user (they can't be, since they don't have any).
-		 */
-		const affectedSites = sites
-			.filter( ( site ) => {
-				if ( ! site?.plan?.is_free ) {
-					return site.capabilities.manage_options;
-				}
-				if ( site?.products && site?.products?.length > 0 ) {
-					return site.capabilities.manage_options;
-				}
-				return false;
-			} )
-			.map( ( site ) => site.slug );
-
-		if ( ! affectedSites.length ) {
-			return;
-		}
-
-		let affectedSitesString = '';
-
-		if ( affectedSites.length === 1 ) {
-			affectedSitesString = affectedSites[ 0 ];
-		} else {
-			const allButLast = affectedSites.slice( 0, -1 ).join( ', ' );
-			const translatedAnd = translate( 'and', {
-				comment: 'last conjunction in a list of blognames: (blog1, blog2,) blog3 _and_ blog4',
-			} );
-			const last = affectedSites[ affectedSites.length - 1 ];
-
-			affectedSitesString = allButLast + ' ' + translatedAnd + ' ' + last;
-		}
-
-		return (
-			<Notice status="is-info" showDismiss={ false }>
-				{ translate(
-					'The upgrades for {{strong}}%(affectedSitesString)s{{/strong}} are owned by another site administrator. ' +
-						'These can only be managed by the purchase owners.',
-					{
-						components: { strong: <strong /> },
-						args: {
-							affectedSitesString,
-						},
-					}
-				) }
-			</Notice>
-		);
 	}
 
 	renderConciergeBanner() {
@@ -197,7 +143,7 @@ class PurchasesList extends Component<
 								eventName="calypso_no_purchases_upgrade_nudge_impression"
 								eventProperties={ commonEventProps }
 							/>
-							{ this.renderPurchasesByOtherAdminsNotice() }
+							<PurchasesByOtherAdminsNotice sites={ sites } />
 							<EmptyContent
 								title={ translate( 'Looking to upgrade?' ) }
 								line={ translate(
